Combine options/option-item
- Update types pattern
- Replace cva with tv
