+++ design-toolkit/components/src/index.ts
@@ -151,9 +151,7 @@ export type {
   SubmenuTriggerProps,
 } from './components/menu/types';
 export { Options, OptionsContext, OptionsSection } from './components/options';
-export type { OptionsProps } from './components/options';
-export { OptionsItem } from './components/options-item';
-export type { IOptionsItem, OptionsItemProps } from './components/options-item';
+export type { OptionsProps } from './components/options/types';
 export { Popover, PopoverTrigger } from './components/popover';
 export { PopoverStyles } from './components/popover/styles';
 export type {
